// src/app/models/company.ts
export class Company {
    id?: string;
    name?: string;
    nit?: string;
    contract?: string;

    //nuevos campos
    email?: string;
    phone?: string;
    address?: string;
    bank_name?: string;
    account_type?: string;
    account_number?: string;
    swift_bic?: string;
    owner_name?: string;
}